Add ability to save CLI environment variables to config file.
Provide commands to able to do the same and also set the config file to save to.

Add argparse 'CONFIG' option to able to specify the config file when invoking the cli from the shell. (" -c config_file)/

General flake cleanup and some improvements to internal helper functions.

Ticket: #10615
Ticket: #10620